refactor: move modifyFieldPlotAt to its own file
