+++ src/game-logic/reducers/field.js
@@ -36,27 +36,10 @@ import { incrementPlotContentAge } from './incrementPlotContentAge'
 import { resetWasShoveled } from './resetWasShoveled'
 import { showNotification } from './showNotification'
 import { processSprinklers } from './processSprinklers'
+import { modifyFieldPlotAt } from './modifyFieldPlotAt'
 
 const { FERTILIZE, OBSERVE, SET_SCARECROW, SET_SPRINKLER } = fieldMode
 const { GROWN } = cropLifeStage
-
-/**
- * @param {farmhand.state} state
- * @param {number} x
- * @param {number} y
- * @param {Function(?farmhand.plotContent)} modifierFn
- * @returns {farmhand.state}
- */
-export const modifyFieldPlotAt = (state, x, y, modifierFn) => {
-  const { field } = state
-  const row = [...field[y]]
-  const plotContent = modifierFn(row[x])
-  row[x] = plotContent
-  const modifiedField = [...field]
-  modifiedField[y] = row
-
-  return { ...state, field: modifiedField }
-}
 
 /**
  * @param {farmhand.state} state
